next_sentence/requirements.txt
numpy
pillow
editdistance
torch
torchdata
torchmetrics
hydra-core
wandb
lightning
transformers
sentence-transformers
datasets
loralib
accelerate
matplotlib
seaborn
nltk
spacy